[M0] Fixed a critical error when creating a locres file.
In fact, when writing the offset to the data. `LocresHelper.cs` line => [87]
Previously, I recorded only the first 2 bytes as offset instead of all 4 bytes, which caused the "Dimension error" error.

Thanks to @koudy-koder for bringing this error to my attention.
I'm ADHD guy, so this maybe perhaps such errors are left somewhere else, sorry, and thank for using my tool!