[M0] New parsing system
A lot of code refactoring, the functionality of parsing files from the unpacked Unreal archive was removed, and a parser via CUE4Parse came to replace it.
This is more reliable, faster, more convenient, and will provide more data in the future when working with ZenLoader and DataTable.

To work with unpacked resources, use the Legacy branch, or its forks, and releases up to 1.0.6.0. Thank you all, this is a great day bringing great changes.